+++ include/ignition/sensors/Sensor.hh
@@ -54,6 +54,11 @@ namespace ignition
       /// \brief destructor
       public: virtual ~Sensor();
 
+      /// \brief Load the sensor based on data from an sdf::Sensor object.
+      /// \param[in] _sdf SDF <sensor> or <plugin> inside of <sensor>
+      /// \return true if loading was successful
+      public: virtual bool Load(const sdf::Sensor &_sdf);
+
       /// \brief Load the sensor with SDF parameters.
       /// \param[in] _sdf SDF <sensor> or <plugin> inside of <sensor>
       /// \return true if loading was successful