+++ src/Sensor.cc
@@ -25,8 +25,8 @@ using namespace ignition::sensors;
 
 class ignition::sensors::SensorPrivate
 {
-  /// \brief Populates fields from a <sensor> element
-  public: bool PopulateFromSDF(sdf::ElementPtr _sdf);
+  /// \brief Populates fields from a <sensor> DOM
+  public: bool PopulateFromSDF(const sdf::Sensor &_sdf);
 
   /// \brief id given to sensor when constructed
   public: SensorId id;
@@ -53,13 +53,18 @@ class ignition::sensors::SensorPrivate
   public: ignition::common::Time nextUpdateTime;
 
   public: sdf::ElementPtr sdf = nullptr;
+
+  /// \brief SDF Sensor DOM object.
+  public: sdf::Sensor sdfSensor;
 };
 
 SensorId SensorPrivate::idCounter = 0;
 
 //////////////////////////////////////////////////
-bool SensorPrivate::PopulateFromSDF(sdf::ElementPtr _sdf)
+bool SensorPrivate::PopulateFromSDF(const sdf::Sensor &_sdf)
 {
+  this->sdfSensor = _sdf;
+
   // All SDF code gets auto converted to latest version. This code is
   // written assuming sdformat 1.6 is the latest
 
@@ -74,39 +79,10 @@ bool SensorPrivate::PopulateFromSDF(sdf::ElementPtr _sdf)
   // be outside the scope of this library
 
   // \todo(nkoenig) how to use frame?
-
-  if (!_sdf)
-  {
-    ignerr << "null _sdf\n";
-    return false;
-  }
-
-  if (std::string("plugin") == _sdf->GetName())
-    _sdf = _sdf->GetParent();
-
-  if (std::string("sensor") != _sdf->GetName())
-  {
-    ignerr << "SDF is not a sensor.\n";
-    _sdf->PrintValues("  ");
-    return false;
-  }
-
-  this->name = _sdf->Get<std::string>("name");
-
-  if (_sdf->HasElement("topic"))
-  {
-    this->topic = _sdf->Get<std::string>("topic");
-  }
-
-  if (_sdf->HasElement("pose"))
-  {
-    this->pose = _sdf->Get<ignition::math::Pose3d>("pose");
-  }
-
-  if (_sdf->HasElement("update_rate"))
-  {
-    this->updateRate = (_sdf->Get<double>("update_rate"));
-  }
+  this->name = _sdf.Name();
+  this->topic = _sdf.Topic();
+  this->pose = _sdf.Pose();
+  this->updateRate = _sdf.UpdateRate();
   return true;
 }
 
@@ -129,6 +105,12 @@ Sensor::~Sensor()
 }
 
 //////////////////////////////////////////////////
+bool Sensor::Load(const sdf::Sensor &_sdf)
+{
+  return this->dataPtr->PopulateFromSDF(_sdf);
+}
+
+//////////////////////////////////////////////////
 bool Sensor::Load(sdf::ElementPtr _sdf)
 {
   if (!this->dataPtr->sdf)
@@ -137,7 +119,10 @@ bool Sensor::Load(sdf::ElementPtr _sdf)
   }
   else
     this->dataPtr->sdf->Copy(_sdf);
-  return this->dataPtr->PopulateFromSDF(_sdf);
+
+  sdf::Sensor sdfSensor;
+  sdfSensor.Load(_sdf);
+  return this->dataPtr->PopulateFromSDF(sdfSensor);
 }
 
 //////////////////////////////////////////////////