+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ### Ignition Sensors 2.X.X (201X-XX-XX)
 
+1. Add support for loading a magnetometer sensor from and SDF Sensor DOM
+   object.
+    * [Pull request 53](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-request/53)
+
 1. Add magnetometer
     * [Pull request 47](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-request/47)